support old functionalities

+++ library_app/redis/library.py
@@ -14,27 +14,31 @@ def parse_dict(dict, key):
     return None
 
 
-def set_hash_and_update_set_reference(hash_key, entity, set_prefix, value):
-    v = get_redis().hget(hash_key, entity)
-    if v:
+def set_hash(hash_key, entity, value):
+    oldValue = get_redis().hget(hash_key, entity)
+    if oldValue:
         if not value:
             get_redis().hdel(hash_key, entity)
-        get_redis().srem(set_prefix + str(v, 'utf-8'), hash_key)
+        oldValue = str(oldValue, 'utf-8')
     if value:
         get_redis().hset(hash_key, entity, value)
-        get_redis().sadd(set_prefix + value, hash_key)
+    return oldValue
 
+def update_set_reference(refer_key, set_prefix, oldValue, newValue):
+    if oldValue and newValue:
+        get_redis().smove(set_prefix + oldValue, set_prefix + newValue, refer_key)
+    elif oldValue:
+        get_redis().srem(set_prefix + oldValue, refer_key)
+    elif newValue:
+        get_redis().sadd(set_prefix + newValue, refer_key)
+
+def set_hash_and_update_set_reference(hash_key, entity, set_prefix, value):
+    oldValue = set_hash(hash_key, entity, value)
+    update_set_reference(hash_key, set_prefix, oldValue, value)
 
 class BookProxy:
-    def __init__(self, book_id):
-        self.book_id = book_id
-        self.book_key = 'book:' + str(book_id)
-
-    @staticmethod
-    def add():
-        bookproxy = BookProxy(get_redis().incr('book:count'))
-        get_redis().sadd('book:keys', bookproxy.book_key)
-        return bookproxy
+    def __init__(self, isbn):
+        self.book_key = 'book:' + str(isbn)
 
     @staticmethod
     def key_to_Book(key):
@@ -43,22 +47,23 @@ class BookProxy:
             key = str(key, 'utf-8')
         if len(dict) == 0:
             return None
-        return Book(id=key[6:],
+        return Book(isbn=parse_dict(dict, b'isbn'),
                     title=parse_dict(dict, b'title'),
                     author=parse_dict(dict, b'author'),
-                    isbn=parse_dict(dict, b'isbn'),
-                    page_num=parse_dict(dict, b'page_num'),
-                    checkoutby=parse_dict(dict, b'checkoutby'))
+                    page_num=parse_dict(dict, b'page_num'))
 
     @staticmethod
-    def getBooks(book_keys):
+    def get_books(book_keys):
         return [BookProxy.key_to_Book(key) for key in book_keys]
+
+    def add(self):
+        get_redis().sadd('book:keys', self.book_key)
 
     def fetch(self):
         return BookProxy.key_to_Book(self.book_key)
 
     def exists(self):
-        return get_redis().sismember('book:keys', self.book_key)
+        return self.book_key in get_redis()
 
     def edit(self, book):
         if book.title:
@@ -69,8 +74,8 @@ class BookProxy:
             self.set_isbn(book.isbn)
         if book.page_num:
             self.set_page_num(book.page_num)
-        if book.checkoutby:
-            self.set_checkoutby(book.checkoutby)
+        if book.quantity:
+            self.set_quantity(book.quantity)
 
     def delete(self):
         get_redis().delete(self.book_key)
@@ -79,7 +84,7 @@ class BookProxy:
         self.set_author(None)
         self.set_isbn(None)
         self.set_page_num(None)
-        self.set_checkoutby(None)
+        get_redis().delete(self.book_key)
         return False
 
     def set_title(self, title):
@@ -91,16 +96,30 @@ class BookProxy:
     def set_isbn(self, isbn):
         set_hash_and_update_set_reference(self.book_key, 'isbn', 'book:isbn-', isbn)
 
+    def set_quantity(self, quantity):
+        if not type(quantity) == int:
+            raise Exception('quantity must be integer :' + str(quantity))
+        set_hash(self.book_key, 'quantity', quantity)
+
+    def get_quantity(self):
+        return int(get_redis().hget(self.book_key, 'quantity'))
+
     def set_page_num(self, page_num):
-        if page_num:
-            get_redis().hset(self.book_key, 'page_num', page_num)
+        set_hash(self.book_key, 'page_num', page_num)
 
-    def get_checkoutby(self):
-        return get_redis().hget(self.book_key, 'checkoutby')
+    def get_borrower_num(self):
+        return get_redis().scard('book:checkoutby-'+self.book_key)
 
-    def set_checkoutby(self, checkoutby):
-        set_hash_and_update_set_reference(self.book_key, 'checkoutby', 'book:checkoutby-', checkoutby)
+    def is_borrower(self, borrowerProxy):
+        return get_redis().sismember('book:checkoutby-'+self.book_key, borrowerProxy.borrower_key)
 
+    def add_borrower(self, borrowerProxy):
+        get_redis().sadd('book:checkoutby-'+self.book_key, borrowerProxy.borrower_key)
+        get_redis().sadd('borrower:checkoutby-'+borrowerProxy.borrower_key, self.book_key)
+
+    def remove_borrower(self, borrowerProxy):
+        get_redis().srem('book:checkoutby-'+self.book_key, borrowerProxy.borrower_key)
+        get_redis().srem('borrower:checkoutby-'+borrowerProxy.borrower_key, self.book_key)
 
 class BorrowerProxy:
     def __init__(self, username):
@@ -135,9 +154,9 @@ class BorrowerProxy:
             self.set_phone(borrower.phone)
 
     def delete(self):
-        get_redis().delete(self.borrower_key)
         self.set_name(None)
         self.set_phone(None)
+        get_redis().delete(self.borrower_key)
 
     def set_name(self, name):
         set_hash_and_update_set_reference(self.borrower_key, 'name', 'borrower:name-', name)
@@ -153,9 +172,12 @@ class RedisLibrary:
 
     def add_book(self, book):
         # book:count stored the largets book id that can exist
-        bookproxy = BookProxy.add()
+        bookproxy = BookProxy(book.isbn)
+        if bookproxy.exists():
+            return False, 'book_exist_already'
+        bookproxy.add()
         bookproxy.edit(book)
-        book.id = bookproxy.book_id
+        return True
 
     def get_book(self, book_id):
         return BookProxy(book_id).fetch()
@@ -164,8 +186,10 @@ class RedisLibrary:
         return False if book does not exists
     '''
 
-    def delete_book(self, book_id):
-        proxy = BookProxy(book_id)
+    def delete_book(self, isbn):
+        if type(isbn) != str:
+            raise Exception('isbn needs to be str')
+        proxy = BookProxy(isbn)
         if not proxy.exists():
             return False
         proxy.delete()
@@ -175,37 +199,38 @@ class RedisLibrary:
         return False if book does not exists
     '''
 
-    def edit_book(self, book_id, book):
-        proxy = BookProxy(book_id)
+    def edit_book(self, isbn, book):
+        proxy = BookProxy(isbn)
         if not proxy.exists():
             return False
         proxy.edit(book)
         return True
 
     def search_by_title(self, title):
-        return BookProxy.getBooks(get_redis().smembers('book:title-' + title))
+        return BookProxy.get_books(get_redis().smembers('book:title-' + title))
 
     def search_by_author(self, author):
-        return BookProxy.getBooks(get_redis().smembers('book:author-' + author))
+        return BookProxy.get_books(get_redis().smembers('book:author-' + author))
 
     def search_by_isbn(self, isbn):
-        return BookProxy.getBooks(get_redis().smembers('book:isbn-' + isbn))
+        return BookProxy.get_books(get_redis().smembers('book:isbn-' + isbn))
 
     def sort_by_title(self):  # return all books
-        return BookProxy.getBooks(get_redis().sort('book:keys', by='*->title', alpha=True))
+        return BookProxy.get_books(get_redis().sort('book:keys', by='*->title', alpha=True))
 
     def sort_by_author(self):  # return all books
-        return BookProxy.getBooks(get_redis().sort('book:keys', by='*->author', alpha=True))
+        return BookProxy.get_books(get_redis().sort('book:keys', by='*->author', alpha=True))
 
     def sort_by_isbn(self):  # return all books
-        return BookProxy.getBooks(get_redis().sort('book:keys', by='*->isbn', alpha=True))
+        return BookProxy.get_books(get_redis().sort('book:keys', by='*->isbn', alpha=True))
 
     def sort_by_page_num(self):  # return all books
-        return BookProxy.getBooks(get_redis().sort('book:keys', by='*->page_num', alpha=True))
+        return BookProxy.get_books(get_redis().sort('book:keys', by='*->page_num', alpha=True))
 
     '''
         return False if user already existed
     '''
+
     def add_borrower(self, borrower):
         proxy = BorrowerProxy(borrower.username)
         if proxy.exists():
@@ -236,36 +261,43 @@ class RedisLibrary:
     '''
         return False if already borrowed by someone else
     '''
-
     def checkout_book(self, username, book_id):
-        if not BorrowerProxy(username).exists():
-            return False, 'username_invalid'
+        borrowerProxy = BorrowerProxy(username)
+        if not borrowerProxy.exists():
+            return False, 'browser_not_exist'
         proxy = BookProxy(book_id)
         if not proxy.exists():
             return False, 'book_not_exist'
-        if proxy.get_checkoutby():
+        # verify availability
+        if proxy.is_borrower(borrowerProxy):
             return False, 'book_already_borrowed'
-        proxy.set_checkoutby(username)
+        if proxy.get_borrower_num() >= proxy.get_quantity():
+            return False, 'book_not_available'
+        proxy.add_borrower(borrowerProxy)
         return True
 
     def return_book(self, username, book_id):
-        if not BorrowerProxy(username).exists():
-            return False, 'username_invalid'
+        borrowerProxy = BorrowerProxy(username)
+        if not borrowerProxy.exists():
+            return False, 'browser_not_exist'
         proxy = BookProxy(book_id)
         if not proxy.exists():
             return False, 'book_not_exist'
-        if not proxy.get_checkoutby():
-            return False, 'book_not_checkedout'
-        proxy.set_checkoutby(None)
+        # verify indeed borrowed
+        if not proxy.is_borrower(borrowerProxy):
+            return False, 'book_not_borrowed'
+        proxy.remove_borrower(borrowerProxy)
         return True
 
-    def get_book_checkedoutby(self, checkoutby):
-        return BookProxy.getBooks(get_redis().smembers('book:checkoutby-' + checkoutby))
-
-    def get_borrower_has(self, book_id):
-        book = BookProxy(book_id).fetch()
-        if not book:
+    def get_book_borrowers(self, isbn):
+        proxy = BookProxy(isbn)
+        if not proxy.exists():
             return None, 'book_not_exist'
-        if book.checkoutby:
-            return BorrowerProxy(book.checkoutby).fetch()
-        return None, 'book_not_checkedout'
+        return BorrowerProxy.get_borrowers(get_redis().smembers('book:checkoutby-' + proxy.book_key))
+
+    def get_borrowed_books(self, username):
+        proxy = BorrowerProxy(username)
+        if not proxy.exists():
+            return None, 'browser_not_exist'
+        return BookProxy.get_books(get_redis().smembers('borrower:checkoutby-' + proxy.borrower_key))
+
